LibIIO

Documentation for LibIIO.

LibIIO.Attr

LibIIO.Buffer

LibIIO.Buffer

LibIIO.Channel

LibIIO.Channel

LibIIO.ChannelAttr

LibIIO.ChannelAttr

LibIIO.Context

LibIIO.Context

LibIIO.Device

LibIIO.Device

LibIIO.DeviceAttr

LibIIO.DeviceBufferAttr

LibIIO.DeviceDebugAttr

LibIIO.LocalContext

LibIIO.LocalContext

LibIIO.NetworkContext

LibIIO.NetworkContext

LibIIO.Trigger

LibIIO.Trigger

LibIIO.XMLContext

LibIIO.XMLContext

Base.length

Base.read

Base.read

Base.read

Base.write

Base.write

Base.write

LibIIO.attrs

LibIIO.attrs

LibIIO.attrs

LibIIO.buffer_attrs

LibIIO.cancel

LibIIO.channels

LibIIO.clone

LibIIO.context

LibIIO.data_format

LibIIO.debug_attrs

LibIIO.description

LibIIO.device

LibIIO.device

LibIIO.devices

LibIIO.enabled

LibIIO.enabled!

LibIIO.filename

LibIIO.find_channel

LibIIO.find_device

LibIIO.frequency

LibIIO.frequency!

LibIIO.id

LibIIO.id

LibIIO.index

LibIIO.label

LibIIO.modifier

LibIIO.name

LibIIO.name

LibIIO.name

LibIIO.output

LibIIO.poll_fd

LibIIO.push

LibIIO.refill

LibIIO.reg_read

LibIIO.reg_write

LibIIO.sample_size

LibIIO.scan_contexts

LibIIO.scan_element

LibIIO.set_blocking_mode

LibIIO.set_kernel_buffers_count

LibIIO.set_timeout

LibIIO.step

LibIIO.trigger

LibIIO.trigger!

LibIIO.type

LibIIO.version

LibIIO.xml

LibIIO.Attr — Type

Abstract super-type for all Attributes-like types

LibIIO.Buffer — Type

This class is used for all I/O operations of buffer capable devices.

LibIIO.Buffer — Method

Buffer(device::T, samples_count, cyclic::Bool = false) where {T <: AbstractDeviceOrTrigger}

Initializes a new Buffer instance.

Parameters

device::AbstractDeviceOrTrigger : A device instance (either Device or Trigger to which the buffer belongs

samples_count : The size of the buffer in samples

cyclic : If set to true, the buffer is circular

LibIIO.Channel — Type

Channel{T <: AbstractDeviceOrTrigger} <: AbstractChannel

Represents a channel of an IIO device.

LibIIO.Channel — Method

Channel(dev::AbstractDeviceOrTrigger, channel::Ptr{iio_channel})

Initializes a new instance of the Channel type.

Parameters

dev::AbstractDeviceOrTrigger : The parent device handle (Device or Trigger)

channel::Ptr{iio_channel} : A valid pointer to an IIO channel.

Returns

A new instance of this type

LibIIO.ChannelAttr — Type

ChannelAttr <: Attr

Represents an attribute of a IIO channel

LibIIO.ChannelAttr — Method

ChannelAttr(chn::Ptr{iio_channel}, name::String)

Initializes a new instance of a ChannelAttr.

Parameters

chn::Ptr{iio_channel} : A valid pointer to an iio_channel

name::String : The channel attribute's name

Returns

A new instance of this type

LibIIO.Context — Type

Context(context = nothing)

Initializes a new Context using the local or the network backend of the IIO library.

This function will create a network context if the IIOD_REMOTE environment variable is set to the hostname where the IIOD server runs. If set to an empty string, the server will be discovered using ZeroConf. If the environment variable is not set, a local context will be created instead.

LibIIO.Context — Type

Contains the representation of an IIO context.

LibIIO.Device — Type

Contains the representation of an IIO device.

LibIIO.Device — Method

Device(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Device instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this device

LibIIO.DeviceAttr — Type

DeviceAttr <: Attr

Represents an attribute of an IIO device

LibIIO.DeviceBufferAttr — Type

DeviceAttr <: Attr

Represents a buffer attribute of an IIO device

LibIIO.DeviceDebugAttr — Type

DeviceAttr <: Attr

Represents a debug attribute of an IIO device

LibIIO.LocalContext — Type

Local IIO Context.

LibIIO.LocalContext — Method

LocalContext([ctx::Ptr{iio_context}])

Initializes a new LocalContext using the local backend if the IIO library. Can be constructed from an existing pointer to an iio_context.

LibIIO.NetworkContext — Type

XMLContext(xml_file)

Initializes a new NetworkContext using the network backend if the IIO library.

Parameters

hostname : Hostname, IPv4 or IPv6 address where the IIO Daemon is running

LibIIO.NetworkContext — Type

Network IIO context

LibIIO.Trigger — Type

Contains the representation of an IIO device that can act as a trigger.

LibIIO.Trigger — Method

Trigger(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Trigger instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this trigger

LibIIO.XMLContext — Type

XML IIO Context.

LibIIO.XMLContext — Method

XMLContext(xml_file)

Initializes a new XMLContext using the XML backend if the IIO library.

Parameters

xmlfile : Filename of the XML file to build the context from.

Base.length — Method

length(buf::Buffer)

Size of this buffer, in bytes.

Base.read — Function

read(chn::Channel, buf::AbstractBuffer, raw=false)

Extract the samples corresponding to this channel from the given buffer.

Parameters

chn::Channel : The channel instance

buf::AbstractBuffer : A buffer instance

raw::Bool: If set to true, the samples are not converted from their native format to their host format

Returns

A Vector{Cuchar} containing the samples for this channel

Base.read — Method

read(buf::Buffer)

Retrieves the samples contained inside the Buffer.

Returns

A Vector{Cuchar} containing the samples

Base.read — Method

read(::Attr)

The current value of this attribute (as string).

Base.write — Method

write(::Attr, value)

Sets the value of the attribute to the passed value.

Base.write — Method

write(buf::Buffer, data::Vector{T}) where {T}

Copy the given vector of samples into the buffer

Parameters

data : The data vector containing the samples to copy

Returns

The number of bytes written into the buffer

Base.write — Method

read(chn::Channel, buf::AbstractBuffer, data::Vector{T}, raw=false) where {T}

Write the specified vector of samples into the buffer the specified channel.

Parameters

chn::Channel : The channel instance

buf::AbstractBuffer : A buffer instance

data::Vector{T} : A data vector containing the samples to copy

raw::Bool: If set to true, the samples are not converted from their host format to their native format

Returns

The number of bytes written

LibIIO.attrs — Method

attrs(ctx::AbstractContext)

Lost of context-specific attributes.

LibIIO.attrs — Method

attrs(d::AbstractDeviceOrTrigger)

List of attributes for the IIO device.

LibIIO.attrs — Method

attrs(chn::Channel)

List of attributes for the given channel.

LibIIO.buffer_attrs — Method

buffer_attrs(d::AbstractDeviceOrTrigger)

List of buffer attributes for the IIO device.

LibIIO.cancel — Method

cancel(buf::Buffer)

Cancel the current buffer.

LibIIO.channels — Method

channels(d::AbstractDeviceOrTrigger)

List of channels available with this IIO device.

LibIIO.clone — Method

clone(ctx::T) where {T <: AbstractContext}

Clones the IIO context.

LibIIO.context — Method

context(d::Device)

Context for the device.

LibIIO.data_format — Method

data_format(chn::Channel)::iio_data_format

Returns the channel data format as a C-struct. See iio_data_format.

LibIIO.debug_attrs — Method

debug_attrs(d::AbstractDeviceOrTrigger)

List of debug attributes for the IIO device.

LibIIO.description — Method

description(ctx::AbstractContext)

Description of this IIO context.

LibIIO.device — Method

device(buf::Buffer)

Device for the buffer.

LibIIO.device — Method

device(chn::Channel)

Retrieves the corresponding AbstractDeviceOrTrigger for this channel.

LibIIO.devices — Method

devices(ctx::AbstractContext)

List of devices contained in this context.

LibIIO.enabled! — Method

enabled!(chn::Channel, state)

Sets the channel state to enabled if true, disabled otherwise.

LibIIO.enabled — Method

enabled(chn::Channel)

Returns true if the channel is enabled, false otherwise.

LibIIO.filename — Method

filename(:Attr)

The filename in sysfs to which this attribute is bound.

LibIIO.find_channel — Function

find_channel(d::AbstractDeviceOrTrigger, name_or_id, is_output = false)

Find an IIO channel by its name or ID.

Parameters

d : The device instance

name_or_id : The name or ID of the channel to find

is_output : Set to true to search for an output channel

Returns

The IIO channel as Channel

LibIIO.find_device — Method

find_device(ctx::AbstractContext, name_or_id_or_label)

Find an IIO device by its name, ID or label.

LibIIO.frequency! — Method

frequency!(t::Trigger, value)

Set the trigger rate.

LibIIO.frequency — Method

frequency(t::Trigger)

Configured frequency (in Hz) of the trigger.

LibIIO.id — Method

id(d::AbstractDeviceOrTrigger)

An identifier of the device, only valid in this IIO context.

LibIIO.id — Method

id(chn::Channel)

An identifier of this channel. Note that it is possible that two channels have the same ID, if one is an input channel and the other is an output channel.

LibIIO.index — Method

index(chn::Channel)

Returns the index of the channel.

LibIIO.label — Method

label(d::AbstractDeviceOrTrigger)

The label of the device.

LibIIO.modifier — Method

modifier(chn::Channel)::iio_modifier

Returns the channel modifier as an enum value.

LibIIO.name — Method

name(ctx::AbstractContext)

Name of the IIO context.

LibIIO.name — Method

name(d::AbstractDeviceOrTrigger)

The name of the device.

LibIIO.name — Method

name(:Attr)

LibIIO.output — Method

output(chn::Channel)

Contains true if the channel is an output channel, false otherwise.

LibIIO.poll_fd — Method

poll_fd(buf::Buffer)

Poll_fd for the buffer.

LibIIO.push — Method

push(buf::Buffer)

Submit the samples contained in this buffer to the hardware.

Parameters

samples_count (optional) : The number of samples to submit, default = full buffer

LibIIO.refill — Method

refill(buf::Buffer)

Fetch a new set of samples from the hardware.

LibIIO.reg_read — Method

reg_read(d::AbstractDeviceOrTrigger, reg)

Read the content of a register of this device.

Parameters

d : The device instance

reg : The register address

Returns

The value of the register

LibIIO.reg_write — Method

reg_write(d::AbstractDeviceOrTrigger, reg, value)

Set a valie to one register of the device.

Parameters

d : The device instance

reg : The register address

value ; The value that will be used for this register

LibIIO.sample_size — Method

sample_size(d::AbstractDeviceOrTrigger)

Sample size of the device. The sample size varies each time channels get enabled or disabled.

LibIIO.scan_contexts — Method

scan_contexts()

Scan Context.

LibIIO.scan_element — Method

scan_element(chn::Channel)

Contains true if the channel is a scan element, false otherwise.

LibIIO.set_blocking_mode — Method

set_blocking_mode(buf::Buffer, blocking)

Set the buffer's blocking mode.

Parameters:

blocking : True if in blocking mode else false.

LibIIO.set_kernel_buffers_count — Method

set_kernel_buffers_count(d::AbstractDeviceOrTrigger, count)

Set the number of kernel buffers to use with the specified device

Parameters

d : The device instance

count : The number of kernel buffers

LibIIO.set_timeout — Method

set_timeout(ctx::AbstractContext, timeout)

Set a timeout for I/O operations.

Parameters

value : The timeout value, in milliseconds

LibIIO.step — Method

step(buf::Buffer)

Step size for the buffer.

LibIIO.trigger! — Method

trigger!(d::Device, trigger::Trigger)

Sets the configured trigger for this IIO device.

LibIIO.trigger — Method

trigger(d::Device)

Returns the configured trigger for this IIO device, if present in the current context.

LibIIO.type — Method

type(chn::Channel)::iio_chan_type

Returns the channel type as an enum value.

LibIIO.version — Method

version(ctx::AbstractContext)

Version of the context backed, as Tuple(UInt, UInt, String)`

LibIIO.xml — Method

xml(ctx::AbstractContext)

XML representation of the IIO context.