LibIIO.Buffer

LibIIO.Buffer

Base.length

Base.read

Base.write

LibIIO.cancel

LibIIO.device

LibIIO.poll_fd

LibIIO.push

LibIIO.refill

LibIIO.set_blocking_mode

LibIIO.step

LibIIO.Buffer — Type

This class is used for all I/O operations of buffer capable devices.

LibIIO.Buffer — Method

Buffer(device::T, samples_count, cyclic::Bool = false) where {T <: AbstractDeviceOrTrigger}

Initializes a new Buffer instance.

Parameters

device::AbstractDeviceOrTrigger : A device instance (either Device or Trigger to which the buffer belongs

samples_count : The size of the buffer in samples

cyclic : If set to true, the buffer is circular

Base.length — Method

length(buf::Buffer)

Size of this buffer, in bytes.

Base.read — Method

read(buf::Buffer)

Retrieves the samples contained inside the Buffer.

Returns

A Vector{Cuchar} containing the samples

Base.write — Method

write(buf::Buffer, data::Vector{T}) where {T}

Copy the given vector of samples into the buffer

Parameters

data : The data vector containing the samples to copy

Returns

The number of bytes written into the buffer

LibIIO.cancel — Method

cancel(buf::Buffer)

Cancel the current buffer.

LibIIO.device — Method

device(buf::Buffer)

Device for the buffer.

LibIIO.poll_fd — Method

poll_fd(buf::Buffer)

Poll_fd for the buffer.

LibIIO.push — Method

push(buf::Buffer [, samples_count])

Submit the samples contained in this buffer to the hardware.

Parameters

buf::Buffer : The buffer struct

samples_count (optional) : The number of samples to submit, default = full buffer

LibIIO.refill — Method

refill(buf::Buffer)

Fetch a new set of samples from the hardware.

LibIIO.set_blocking_mode — Method

set_blocking_mode(buf::Buffer, blocking)

Set the buffer's blocking mode.

Parameters:

blocking : True if in blocking mode else false.

LibIIO.step — Method

step(buf::Buffer)

Step size for the buffer.