Index

LibIIO.CLibIIO.iio_channel_attr_get_filename

LibIIO.CLibIIO.iio_channel_attr_read

LibIIO.CLibIIO.iio_channel_attr_read_all

LibIIO.CLibIIO.iio_channel_attr_read_bool

LibIIO.CLibIIO.iio_channel_attr_read_double

LibIIO.CLibIIO.iio_channel_attr_read_longlong

LibIIO.CLibIIO.iio_channel_attr_write

LibIIO.CLibIIO.iio_channel_attr_write_all

LibIIO.CLibIIO.iio_channel_attr_write_bool

LibIIO.CLibIIO.iio_channel_attr_write_double

LibIIO.CLibIIO.iio_channel_attr_write_longlong

LibIIO.CLibIIO.iio_channel_attr_write_raw

LibIIO.CLibIIO.iio_channel_disable

LibIIO.CLibIIO.iio_channel_enable

LibIIO.CLibIIO.iio_channel_find_attr

LibIIO.CLibIIO.iio_channel_get_attr

LibIIO.CLibIIO.iio_channel_get_attrs_count

LibIIO.CLibIIO.iio_channel_get_data

LibIIO.CLibIIO.iio_channel_get_device

LibIIO.CLibIIO.iio_channel_get_id

LibIIO.CLibIIO.iio_channel_get_modifier

LibIIO.CLibIIO.iio_channel_get_name

LibIIO.CLibIIO.iio_channel_get_type

LibIIO.CLibIIO.iio_channel_is_enabled

LibIIO.CLibIIO.iio_channel_is_output

LibIIO.CLibIIO.iio_channel_is_scan_element

LibIIO.CLibIIO.iio_channel_read

LibIIO.CLibIIO.iio_channel_read_raw

LibIIO.CLibIIO.iio_channel_set_data

LibIIO.CLibIIO.iio_channel_write

LibIIO.CLibIIO.iio_channel_write_raw

Documentation

LibIIO.CLibIIO.iio_channel_get_device — Function

iio_channel_get_device(chn)

Retrieve a pointer to the iio_device structure.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

A pointer to a iio_device structure

See libiio

LibIIO.CLibIIO.iio_channel_get_id — Function

iio_channel_get_id(chn)

Retrieve the channel ID (e.g. voltage0)

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

A string with the channel ID

See libiio

LibIIO.CLibIIO.iio_channel_get_name — Function

iio_channel_get_name(chn)

Retrieve the channel name (e.g. vccint)

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

A string with the channel name

Note

If the channel has no name, an empty string is returned.

See libiio

LibIIO.CLibIIO.iio_channel_is_output — Function

iio_channel_is_output(chn)

Return true if the given channel is an output channel

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

True if the channel is an output channel, false otherweise

See libiio

LibIIO.CLibIIO.iio_channel_is_scan_element — Function

iio_channel_is_scan_element(chn)

Return true of the given channel is a scan element.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

True if the channel is a scan element, false otherweise

Note

A channel that is a scan element is a channel that can generate samples (for an input channel) or receive samples (for an output channel) after being enabled.

See libiio

LibIIO.CLibIIO.iio_channel_get_attrs_count — Function

iio_channel_get_attrs_count(chn)

Enumerate the channel-specific attributes of the given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

The number of channel-specific attributes found

See libiio

LibIIO.CLibIIO.iio_channel_get_attr — Function

iio_channel_get_attr(chn, index)

Get the channel-specific attribute present at the given index.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

index::Cuint : The index corresponding to the attribute

Returns

On success, a string containing the attribute name

If the index is invalid, an empty string will be returned

See libiio

LibIIO.CLibIIO.iio_channel_find_attr — Function

iio_channel_find_attr(chn, name)

Try to find a channel-specific attribute by its name.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

name::String : A string corresponding to the name of the attribute

Returns

On succes, a string containing the attribute name

If the name does not correspond to any know atribute of the channel, an empty string is returned

Note

This function is useful to detect the presence of an attribute.

See libiio

LibIIO.CLibIIO.iio_channel_attr_get_filename — Function

iio_channel_attr_get_filename(chn, attr)

Retrieve the filename of an attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String ; A string corresponding to the name of the attribute

Returns

On success, a string with the filename

If the attribute is unknown, an empty string is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_read — Function

iio_channel_attr_read(dev, attr)

Read the content of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

Returns

On success, a Vector of Tuples of attribute index and value if attr is an empty string, otherwise a string with the attribute value.

On error, a negative errno code is returned

Note

By an empty string as the "attr" argument to iio_channel_attr_read, it is now possible to read all of the attributes of a device.

See libiio

LibIIO.CLibIIO.iio_channel_attr_read_all — Function

iio_channel_attr_read_all(dev, cb, data)

Read the content of all channel-specific attributes.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

cb::Ptr{Cvoid} : A pointer to a callback function

data::Ptr{Cvoid} : A pointer that will be passed to the callback function

Returns

On success, 0 is returned

On error, a negative errno code is returned

Note

This function is especially useful when used with the network backend, as all the channel-specific attributes are read in one single command.

See libiio

LibIIO.CLibIIO.iio_channel_attr_read_bool — Function

iio_channel_attr_read_bool(dev, attr)

Read the content of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

Returns

On success, 0 and the bool value is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_read_longlong — Function

iio_channel_attr_read_longlong(dev, attr)

Read the content of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

Returns

On success, 0 and the Clonglong value is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_read_double — Function

iio_channel_attr_read_double(dev, attr)

Read the content of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

Returns

On success, 0 and the Cdouble value is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_write — Function

iio_channel_attr_write(dev, attr, src)

Set the value of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

src::String A string to set the attribute to

Returns

On success, the number of bytes written

On error, a negative errno code is returned

Note

By passing C_NULL as the "attr" argument to iio_channel_attr_write, it is now possible to write all of the attributes of a device.

The buffer must contain one block of data per attribute of the device, by the order they appear in the iio_device structure.

The first four bytes of one block correspond to a 32-bit signed value in network order. If negative, the attribute is not written; if positive, it corresponds to the length of the data to write. In that case, the rest of the block must contain the data.

See libiio

LibIIO.CLibIIO.iio_channel_attr_write_raw — Function

iio_channel_attr_write_raw(dev, attr, src, len)

Set the value of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

src::Ptr{Cvoid} : A pointer to the data to be written

len::Csize_t : The number of bytes that should be written

Returns

On success, the number of bytes written

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_write_all — Function

iio_channel_attr_write_all(dev, cb, data)

Set the values of all channel-specifc attributes.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

cb::Ptr{Cvoid} : A pointer to a callback function

data::Ptr{Cvoid} : A pointer that will be passed to the callback function

Returns

On success, 0 is returned

On error, a negative errno code is returned

Note

This function is especially useful when used with the network backend, as all the chsnnel-specific attributes are written in one single command.

See libiio

LibIIO.CLibIIO.iio_channel_attr_write_bool — Function

iio_channel_attr_write_bool(dev, attr, val)

Set the value of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

val::Bool : A bool value to set the attribute to

Returns

On success, 0 is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_write_longlong — Function

iio_channel_attr_write_longlong(dev, attr, val)

Set the value of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

val::Clonglong : A Clonglong value to set the attribute to

Returns

On success, 0 is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_attr_write_double — Function

iio_channel_attr_write_double(dev, attr, val)

Set the value of the given channel-specific attribute.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

attr::String : A string corresponding to the name of the attribute

val::Cdouble : A Cdouble value to set the attribute to

Returns

On success, 0 is returned

On error, a negative errno code is returned

See libiio

LibIIO.CLibIIO.iio_channel_enable — Function

iio_channel_enable(chn)

Enable the given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Note

Before creating an iio_buffer structure with iio_device_create_buffer, it is required to enable at least one channel of the device to read from.

See libiio

LibIIO.CLibIIO.iio_channel_disable — Function

iio_channel_disable(chn)

Disable the given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

See libiio

LibIIO.CLibIIO.iio_channel_is_enabled — Function

iio_channel_is_enabled(chn)

Returns true of the channel is enabled.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

True if the channel is enabled, false otherwise

See libiio

LibIIO.CLibIIO.iio_channel_read_raw — Function

iio_channel_read_raw(chn, buffer, dst)

Demultiplex the samples of a given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

buffer::Ptr{iio_buffer} : A pointer to an iio_buffer structure

dst::Vector{Cuchar} : A destination memory area where the demultiplexed data will be stored

Returns

The size of the demultiplexed data, in bytes

See libiio

LibIIO.CLibIIO.iio_channel_read — Function

iio_channel_read(chn, buffer, dst)

Demultiplex and convert the samples of a given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

buffer::Ptr{iio_buffer} : A pointer to an iio_buffer structure

dst::Vector{Cuchar} : A destination memory area where the converted data will be stored

Returns

The size of the converted data, in bytes

See libiio

LibIIO.CLibIIO.iio_channel_write_raw — Function

iio_channel_write_raw(chn, buffer, dst)

Multiplex the samples of a given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

buffer::Ptr{iio_buffer} : A pointer to an iio_buffer structure

src::Vector{Cuchar} : A source memory area where the sequential data will be read from

Returns

The number of bytes actually multiplexed

See libiio

LibIIO.CLibIIO.iio_channel_write — Function

iio_channel_write(chn, buffer, dst)

Convert and multiplex the samples of a given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

buffer::Ptr{iio_buffer} : A pointer to an iio_buffer structure

src::Vector{Cuchar} : A source memory area where the sequential data will be read from

Returns

The number of bytes actually converted and multiplexed

See libiio

LibIIO.CLibIIO.iio_channel_set_data — Function

iio_channel_set_data(dev, data)

Associate a pointer to an iio_channel structure.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

data::Ptr{Cvoid} : The pointer to be associated

See libiio

LibIIO.CLibIIO.iio_channel_get_data — Function

iio_channel_get_data(dev)

Retrieve a previously associated pointer of an iio_channel structure.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

The pointer previously associated if present, or C_NULL

See libiio

LibIIO.CLibIIO.iio_channel_get_type — Function

iio_channel_get_type(chn)

Get the type of the given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

The type of the channel

See libiio

LibIIO.CLibIIO.iio_channel_get_modifier — Function

iio_channel_get_modifier(chn)

Get the modifier type of the given channel.

Parameters

chn::Ptr{iio_channel} : A pointer to an iio_channel structure

Returns

The modifier type of the channel

See libiio