LibIIO.Channel

LibIIO.Channel

LibIIO.ChannelAttr

LibIIO.ChannelAttr

Base.read

Base.write

LibIIO.attrs

LibIIO.data_format

LibIIO.device

LibIIO.enabled

LibIIO.enabled!

LibIIO.id

LibIIO.index

LibIIO.modifier

LibIIO.output

LibIIO.scan_element

LibIIO.type

LibIIO.Channel — Type

Channel{T <: AbstractDeviceOrTrigger} <: AbstractChannel

Represents a channel of an IIO device.

LibIIO.Channel — Method

Channel(dev::AbstractDeviceOrTrigger, channel::Ptr{iio_channel})

Initializes a new instance of the Channel type.

Parameters

dev::AbstractDeviceOrTrigger : The parent device handle (Device or Trigger)

channel::Ptr{iio_channel} : A valid pointer to an IIO channel.

Returns

A new instance of this type

LibIIO.ChannelAttr — Type

ChannelAttr <: Attr

Represents an attribute of a IIO channel

LibIIO.ChannelAttr — Method

ChannelAttr(chn::Ptr{iio_channel}, name::String)

Initializes a new instance of a ChannelAttr.

Parameters

chn::Ptr{iio_channel} : A valid pointer to an iio_channel

name::String : The channel attribute's name

Returns

A new instance of this type

Base.read — Function

read(chn::Channel, buf::AbstractBuffer, raw=false)

Extract the samples corresponding to this channel from the given buffer.

Parameters

chn::Channel : The channel instance

buf::AbstractBuffer : A buffer instance

raw::Bool: If set to true, the samples are not converted from their native format to their host format

Returns

A Vector{Cuchar} containing the samples for this channel

Base.write — Method

read(chn::Channel, buf::AbstractBuffer, data::Vector{T}, raw=false) where {T}

Write the specified vector of samples into the buffer the specified channel.

Parameters

chn::Channel : The channel instance

buf::AbstractBuffer : A buffer instance

data::Vector{T} : A data vector containing the samples to copy

raw::Bool: If set to true, the samples are not converted from their host format to their native format

Returns

The number of bytes written

LibIIO.attrs — Method

attrs(chn::Channel)

List of attributes for the given channel.

LibIIO.data_format — Method

data_format(chn::Channel)::iio_data_format

Returns the channel data format as a C-struct. See iio_data_format.

LibIIO.device — Method

device(chn::Channel)

Retrieves the corresponding AbstractDeviceOrTrigger for this channel.

LibIIO.enabled! — Method

enabled!(chn::Channel, state)

Sets the channel state to enabled if true, disabled otherwise.

LibIIO.enabled — Method

enabled(chn::Channel)

Returns true if the channel is enabled, false otherwise.

LibIIO.id — Method

id(chn::Channel)

An identifier of this channel. Note that it is possible that two channels have the same ID, if one is an input channel and the other is an output channel.

LibIIO.index — Method

index(chn::Channel)

Returns the index of the channel.

LibIIO.modifier — Method

modifier(chn::Channel)::iio_modifier

Returns the channel modifier as an enum value.

LibIIO.output — Method

output(chn::Channel)

Contains true if the channel is an output channel, false otherwise.

LibIIO.scan_element — Method

scan_element(chn::Channel)

Contains true if the channel is a scan element, false otherwise.

LibIIO.type — Method

type(chn::Channel)::iio_chan_type

Returns the channel type as an enum value.