LibIIO.Buffer

LibIIO.Buffer

LibIIO.CLibIIO.iio_buffer

LibIIO.CLibIIO.iio_chan_type

LibIIO.CLibIIO.iio_channel

LibIIO.CLibIIO.iio_context

LibIIO.CLibIIO.iio_context_info

LibIIO.CLibIIO.iio_data_format

LibIIO.CLibIIO.iio_device

LibIIO.CLibIIO.iio_modifier

LibIIO.Channel

LibIIO.Channel

LibIIO.ChannelAttr

LibIIO.ChannelAttr

LibIIO.Context

LibIIO.Context

LibIIO.Device

LibIIO.Device

LibIIO.DeviceAttr

LibIIO.DeviceBufferAttr

LibIIO.DeviceDebugAttr

LibIIO.LocalContext

LibIIO.LocalContext

LibIIO.NetworkContext

LibIIO.NetworkContext

LibIIO.Trigger

LibIIO.Trigger

LibIIO.XMLContext

LibIIO.XMLContext

Base.length

Base.read

Base.read

Base.write

Base.write

LibIIO.CLibIIO.iio_buffer_cancel

LibIIO.CLibIIO.iio_buffer_destroy

LibIIO.CLibIIO.iio_buffer_end

LibIIO.CLibIIO.iio_buffer_first

LibIIO.CLibIIO.iio_buffer_foreach_sample

LibIIO.CLibIIO.iio_buffer_get_data

LibIIO.CLibIIO.iio_buffer_get_device

LibIIO.CLibIIO.iio_buffer_get_poll_fd

LibIIO.CLibIIO.iio_buffer_push

LibIIO.CLibIIO.iio_buffer_push_partial

LibIIO.CLibIIO.iio_buffer_refill

LibIIO.CLibIIO.iio_buffer_set_blocking_mode

LibIIO.CLibIIO.iio_buffer_set_data

LibIIO.CLibIIO.iio_buffer_start

LibIIO.CLibIIO.iio_buffer_step

LibIIO.CLibIIO.iio_channel_attr_get_filename

LibIIO.CLibIIO.iio_channel_attr_read

LibIIO.CLibIIO.iio_channel_attr_read_all

LibIIO.CLibIIO.iio_channel_attr_read_bool

LibIIO.CLibIIO.iio_channel_attr_read_double

LibIIO.CLibIIO.iio_channel_attr_read_longlong

LibIIO.CLibIIO.iio_channel_attr_write

LibIIO.CLibIIO.iio_channel_attr_write_all

LibIIO.CLibIIO.iio_channel_attr_write_bool

LibIIO.CLibIIO.iio_channel_attr_write_double

LibIIO.CLibIIO.iio_channel_attr_write_longlong

LibIIO.CLibIIO.iio_channel_attr_write_raw

LibIIO.CLibIIO.iio_channel_disable

LibIIO.CLibIIO.iio_channel_enable

LibIIO.CLibIIO.iio_channel_find_attr

LibIIO.CLibIIO.iio_channel_get_attr

LibIIO.CLibIIO.iio_channel_get_attrs_count

LibIIO.CLibIIO.iio_channel_get_data

LibIIO.CLibIIO.iio_channel_get_device

LibIIO.CLibIIO.iio_channel_get_id

LibIIO.CLibIIO.iio_channel_get_modifier

LibIIO.CLibIIO.iio_channel_get_name

LibIIO.CLibIIO.iio_channel_get_type

LibIIO.CLibIIO.iio_channel_is_enabled

LibIIO.CLibIIO.iio_channel_is_output

LibIIO.CLibIIO.iio_channel_is_scan_element

LibIIO.CLibIIO.iio_channel_read

LibIIO.CLibIIO.iio_channel_read_raw

LibIIO.CLibIIO.iio_channel_set_data

LibIIO.CLibIIO.iio_channel_write

LibIIO.CLibIIO.iio_channel_write_raw

LibIIO.CLibIIO.iio_context_clone

LibIIO.CLibIIO.iio_context_destroy

LibIIO.CLibIIO.iio_context_find_device

LibIIO.CLibIIO.iio_context_get_attr

LibIIO.CLibIIO.iio_context_get_attr_value

LibIIO.CLibIIO.iio_context_get_attrs_count

LibIIO.CLibIIO.iio_context_get_description

LibIIO.CLibIIO.iio_context_get_device

LibIIO.CLibIIO.iio_context_get_devices_count

LibIIO.CLibIIO.iio_context_get_name

LibIIO.CLibIIO.iio_context_get_version

LibIIO.CLibIIO.iio_context_get_xml

LibIIO.CLibIIO.iio_context_info_get_description

LibIIO.CLibIIO.iio_context_info_get_uri

LibIIO.CLibIIO.iio_context_info_list_free

LibIIO.CLibIIO.iio_context_set_timeout

LibIIO.CLibIIO.iio_create_context_from_uri

LibIIO.CLibIIO.iio_create_default_context

LibIIO.CLibIIO.iio_create_local_context

LibIIO.CLibIIO.iio_create_network_context

LibIIO.CLibIIO.iio_create_scan_block

LibIIO.CLibIIO.iio_create_scan_context

LibIIO.CLibIIO.iio_create_xml_context

LibIIO.CLibIIO.iio_create_xml_context_mem

LibIIO.CLibIIO.iio_device_attr_read

LibIIO.CLibIIO.iio_device_attr_read_all

LibIIO.CLibIIO.iio_device_attr_read_bool

LibIIO.CLibIIO.iio_device_attr_read_double

LibIIO.CLibIIO.iio_device_attr_read_longlong

LibIIO.CLibIIO.iio_device_attr_write

LibIIO.CLibIIO.iio_device_attr_write_all

LibIIO.CLibIIO.iio_device_attr_write_bool

LibIIO.CLibIIO.iio_device_attr_write_double

LibIIO.CLibIIO.iio_device_attr_write_longlong

LibIIO.CLibIIO.iio_device_attr_write_raw

LibIIO.CLibIIO.iio_device_buffer_attr_read

LibIIO.CLibIIO.iio_device_buffer_attr_read_all

LibIIO.CLibIIO.iio_device_buffer_attr_read_bool

LibIIO.CLibIIO.iio_device_buffer_attr_read_double

LibIIO.CLibIIO.iio_device_buffer_attr_read_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write

LibIIO.CLibIIO.iio_device_buffer_attr_write_all

LibIIO.CLibIIO.iio_device_buffer_attr_write_bool

LibIIO.CLibIIO.iio_device_buffer_attr_write_double

LibIIO.CLibIIO.iio_device_buffer_attr_write_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write_raw

LibIIO.CLibIIO.iio_device_create_buffer

LibIIO.CLibIIO.iio_device_find_attr

LibIIO.CLibIIO.iio_device_find_buffer_attr

LibIIO.CLibIIO.iio_device_find_channel

LibIIO.CLibIIO.iio_device_get_attr

LibIIO.CLibIIO.iio_device_get_attrs_count

LibIIO.CLibIIO.iio_device_get_buffer_attr

LibIIO.CLibIIO.iio_device_get_buffer_attrs_count

LibIIO.CLibIIO.iio_device_get_channel

LibIIO.CLibIIO.iio_device_get_channels_count

LibIIO.CLibIIO.iio_device_get_context

LibIIO.CLibIIO.iio_device_get_data

LibIIO.CLibIIO.iio_device_get_id

LibIIO.CLibIIO.iio_device_get_label

LibIIO.CLibIIO.iio_device_get_name

LibIIO.CLibIIO.iio_device_get_trigger

LibIIO.CLibIIO.iio_device_is_trigger

LibIIO.CLibIIO.iio_device_set_data

LibIIO.CLibIIO.iio_device_set_kernel_buffers_count

LibIIO.CLibIIO.iio_device_set_trigger

LibIIO.CLibIIO.iio_get_backend

LibIIO.CLibIIO.iio_get_backends_count

LibIIO.CLibIIO.iio_has_backend

LibIIO.CLibIIO.iio_library_get_version

LibIIO.CLibIIO.iio_scan_block_destroy

LibIIO.CLibIIO.iio_scan_block_get_info

LibIIO.CLibIIO.iio_scan_block_scan

LibIIO.CLibIIO.iio_scan_context_destroy

LibIIO.CLibIIO.iio_scan_context_get_info_list

LibIIO.CLibIIO.iio_strerror

LibIIO.attrs

LibIIO.attrs

LibIIO.attrs

LibIIO.buffer_attrs

LibIIO.cancel

LibIIO.channels

LibIIO.clone

LibIIO.context

LibIIO.data_format

LibIIO.debug_attrs

LibIIO.description

LibIIO.device

LibIIO.device

LibIIO.devices

LibIIO.enabled

LibIIO.enabled!

LibIIO.find_channel

LibIIO.find_device

LibIIO.frequency

LibIIO.frequency!

LibIIO.id

LibIIO.id

LibIIO.index

LibIIO.label

LibIIO.modifier

LibIIO.name

LibIIO.name

LibIIO.output

LibIIO.poll_fd

LibIIO.push

LibIIO.refill

LibIIO.reg_read

LibIIO.reg_write

LibIIO.sample_size

LibIIO.scan_contexts

LibIIO.scan_element

LibIIO.set_blocking_mode

LibIIO.set_kernel_buffers_count

LibIIO.set_timeout

LibIIO.step

LibIIO.trigger

LibIIO.trigger!

LibIIO.type

LibIIO.version

LibIIO.xml

LibIIO.CLibIIO.iio_buffer — Type

An input or output buffer, used to read or write samples

See libiio

LibIIO.CLibIIO.iio_chan_type — Type

iio_chan_type

IIO channel type

An IIO channel has a type specifying the type of data associated with the channel.

See libiio

LibIIO.CLibIIO.iio_channel — Type

Represents an input or output channel of a device

See libiio

LibIIO.CLibIIO.iio_context — Type

Contains the representation of an IIO context

See libiio

LibIIO.CLibIIO.iio_context_info — Type

The information related to a discovered context

See libiio

LibIIO.CLibIIO.iio_data_format — Type

iio_data_format

Contains the format of a data sample.

Data Fields

length::Cuint : Total length of the sample in bits

bits::Cuint : Length of the valuable data in the samples, in bits

shift:Cuint : Right-shift to apply when converting sample

is_signed::Cuchar : Contains true if the sample is signed

is_fully_defined::Cuchar : Contains true if the sample is fully defoned, sign extended, etc.

is_be::Cuchar : Contains true if the sample is in big-endian format

with_scale::Cuchar : Contains true if the sample should be scaled or converted

scale::Cdouble : Contains the scale to apply if with_scale is set

repeat::Cuint : Numver of times length repeats

See libiio

LibIIO.CLibIIO.iio_device — Type

Represents a device in the IIO context

See libiio

LibIIO.CLibIIO.iio_modifier — Type

iio_modifier

IIO channel type.

An IIO channel has a type specifying the type of data associated with the channel.

See libiio