LibIIO.Context

LibIIO.Context

LibIIO.LocalContext

LibIIO.LocalContext

LibIIO.NetworkContext

LibIIO.NetworkContext

LibIIO.XMLContext

LibIIO.XMLContext

LibIIO.attrs

LibIIO.clone

LibIIO.description

LibIIO.devices

LibIIO.find_device

LibIIO.name

LibIIO.scan_contexts

LibIIO.set_timeout

LibIIO.version

LibIIO.xml

LibIIO.Context — Type

Context(ptr_uri_or_nothing = nothing)

Initializes a new Context using the local or the network backend of the IIO library.

This function will create a network context if the IIOD_REMOTE environment variable is set to the hostname where the IIOD server runs. If set to an empty string, the server will be discovered using ZeroConf. If the environment variable is not set, a local context will be created instead.

Parameters

ptr_uri_or_nothing : Either a Ptr{iio_context}, an URI string (recommended) or nothing to construct the default context (Linux only).

LibIIO.Context — Type

Contains the representation of an IIO context.

LibIIO.LocalContext — Type

Local IIO Context.

LibIIO.LocalContext — Method

LocalContext([ctx::Ptr{iio_context}])

Initializes a new LocalContext using the local backend if the IIO library. Can be constructed from an existing pointer to an iio_context.

LibIIO.NetworkContext — Type

XMLContext(xml_file)

Initializes a new NetworkContext using the network backend if the IIO library.

Parameters

hostname : Hostname, IPv4 or IPv6 address where the IIO Daemon is running

LibIIO.NetworkContext — Type

Network IIO context

LibIIO.XMLContext — Type

XML IIO Context.

LibIIO.XMLContext — Method

XMLContext(xml_file)

Initializes a new XMLContext using the XML backend if the IIO library.

Parameters

xmlfile : Filename of the XML file to build the context from.

LibIIO.attrs — Method

attrs(ctx::AbstractContext)

Lost of context-specific attributes.

LibIIO.clone — Method

clone(ctx::T) where {T <: AbstractContext}

Clones the IIO context.

LibIIO.description — Method

description(ctx::AbstractContext)

Description of this IIO context.

LibIIO.devices — Method

devices(ctx::AbstractContext)

List of devices contained in this context.

LibIIO.find_device — Method

find_device(ctx::AbstractContext, name_or_id_or_label)

Find an IIO device by its name, ID or label.

LibIIO.name — Method

name(ctx::AbstractContext)

Name of the IIO context.

LibIIO.scan_contexts — Method

scan_contexts()

Scan Context.

LibIIO.set_timeout — Method

set_timeout(ctx::AbstractContext, timeout)

Set a timeout for I/O operations.

Parameters

value : The timeout value, in milliseconds

LibIIO.version — Method

version(ctx::AbstractContext)

Version of the context backed, as Tuple(UInt, UInt, String)`

LibIIO.xml — Method

xml(ctx::AbstractContext)

XML representation of the IIO context.