LibIIO.CLibIIO.iio_context_info_get_description — Method

iio_context_info_get_description(info)

Get a description of a discovered context.

Parameters

Ptr{iio_context_info}: A pointer to a iio_context_info structure.

Returns

A String containing the description.

See libiio

LibIIO.CLibIIO.iio_context_info_get_uri — Method

iio_context_info_get_uri(info)

Get the URI of a discovered context.

Parameters

Ptr{iio_context_info}: A pointer to a iio_context_info structure.

Returns

A String containing the URI.

See libiio

LibIIO.CLibIIO.iio_context_info_list_free — Method

iio_context_info_list_free(info)

Free a context info list.

Parameters

info::Ptr{Ptr{iio_context_info}} : A pointer to a iio_context_info pointer.

See libiio

LibIIO.CLibIIO.iio_create_scan_block — Function

iio_create_scan_block(backend, flags = 0)

Parameters

backend::String : A string containing the backend to use for scanning. Can be empty to use all available backends.

flags::Cuint (optional) : Unused for now. Set to 0.

Returns

on success, a pointer to an iio_scan_block structure

on failure, an error is raised

See libiio

LibIIO.CLibIIO.iio_create_scan_context — Function

iio_create_scan_context(backend, flags = 0)

Create a scan context.

Parameters

backend::String : A string containing a comma-separated list of the backend(s) to use for scanning.

flags::Cuint : Unused for now. Set to 0.

Returns

on success, a pointer to a iio_scan_context structure

On failure, an error is raised

NOTE: Libiio version 0.20 and above can handle multiple strings, for instance "local:usb:", "ip:usb:", "local:usb:ip:", and require a colon as the delimiter. Libiio version 0.24 and above prefer a comma instead of colon as the delimiter, and handle specifying backend-specific information. For instance, "local,usb=0456:*" will scan the local backend and limit scans on USB to vendor ID 0x0456, and accept all product IDs. The "usb=0456:b673" string would limit the scan to the device with this particular VID/PID. Both IDs are expected in hexadecimal, no 0x prefix needed.

LibIIO.CLibIIO.iio_scan_block_destroy — Method

iio_scan_block_destroy(blk)

Destroy the given scan block.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

NOTE: After that function, the iioscanblock pointer shall be invalid.

Introduced in version 0.20.

See libiio

LibIIO.CLibIIO.iio_scan_block_get_info — Method

iio_scan_block_get_info(blk, index)

Get the iio_context_info for a particular context.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

index::Cuint : The index corresponding to the context.

Returns

A pointer to the iio_context_info for the context

On success, a pointer to the specified iio_context_info

On failure, an error is raised

LibIIO.CLibIIO.iio_scan_block_scan — Method

iio_scan_block_scan(blk)

Enumerate available contexts via scan block.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

Returns

On success, the number of contexts found.

On failure, a negative error number.

See libiio

LibIIO.CLibIIO.iio_scan_context_destroy — Method

iio_scan_context_destroy(ctx)

Destroy the given scan context.

Parameters

ctx::Ptr{iio_scan_context} : A pointer to an iio_scan_context structure

NOTE: After that function, the iio_scan_context pointer shall be invalid.

See libiio

LibIIO.CLibIIO.iio_scan_context_get_info_list — Method

iio_scan_context_get_info_list(ctx, info)

Enumerate available contexts.

Parameters

ctx::Ptr{iio_scan_context} : A pointer to an iio_scan_context structure

info::Ref{Ptr{Ptr{iio_context_info}}} : A pointer to a list of iio_context_info pointers. The pointed variable will be initialized on success.

Returns

On success, the number of contexts found.

On failure, a negative error number.

See libiio