LibIIO.Buffer

LibIIO.Buffer

LibIIO.CLibIIO.iio_buffer

LibIIO.CLibIIO.iio_chan_type

LibIIO.CLibIIO.iio_channel

LibIIO.CLibIIO.iio_context

LibIIO.CLibIIO.iio_context_info

LibIIO.CLibIIO.iio_data_format

LibIIO.CLibIIO.iio_device

LibIIO.CLibIIO.iio_modifier

LibIIO.Channel

LibIIO.Channel

LibIIO.ChannelAttr

LibIIO.ChannelAttr

LibIIO.Context

LibIIO.Context

LibIIO.Device

LibIIO.Device

LibIIO.DeviceAttr

LibIIO.DeviceBufferAttr

LibIIO.DeviceDebugAttr

LibIIO.LocalContext

LibIIO.LocalContext

LibIIO.NetworkContext

LibIIO.NetworkContext

LibIIO.Trigger

LibIIO.Trigger

LibIIO.XMLContext

LibIIO.XMLContext

Base.length

Base.read

Base.read

Base.write

Base.write

LibIIO.CLibIIO.iio_buffer_cancel

LibIIO.CLibIIO.iio_buffer_destroy

LibIIO.CLibIIO.iio_buffer_end

LibIIO.CLibIIO.iio_buffer_first

LibIIO.CLibIIO.iio_buffer_foreach_sample

LibIIO.CLibIIO.iio_buffer_get_data

LibIIO.CLibIIO.iio_buffer_get_device

LibIIO.CLibIIO.iio_buffer_get_poll_fd

LibIIO.CLibIIO.iio_buffer_push

LibIIO.CLibIIO.iio_buffer_push_partial

LibIIO.CLibIIO.iio_buffer_refill

LibIIO.CLibIIO.iio_buffer_set_blocking_mode

LibIIO.CLibIIO.iio_buffer_set_data

LibIIO.CLibIIO.iio_buffer_start

LibIIO.CLibIIO.iio_buffer_step

LibIIO.CLibIIO.iio_channel_attr_get_filename

LibIIO.CLibIIO.iio_channel_attr_read

LibIIO.CLibIIO.iio_channel_attr_read_all

LibIIO.CLibIIO.iio_channel_attr_read_bool

LibIIO.CLibIIO.iio_channel_attr_read_double

LibIIO.CLibIIO.iio_channel_attr_read_longlong

LibIIO.CLibIIO.iio_channel_attr_write

LibIIO.CLibIIO.iio_channel_attr_write_all

LibIIO.CLibIIO.iio_channel_attr_write_bool

LibIIO.CLibIIO.iio_channel_attr_write_double

LibIIO.CLibIIO.iio_channel_attr_write_longlong

LibIIO.CLibIIO.iio_channel_attr_write_raw

LibIIO.CLibIIO.iio_channel_convert

LibIIO.CLibIIO.iio_channel_convert_inverse

LibIIO.CLibIIO.iio_channel_disable

LibIIO.CLibIIO.iio_channel_enable

LibIIO.CLibIIO.iio_channel_find_attr

LibIIO.CLibIIO.iio_channel_get_attr

LibIIO.CLibIIO.iio_channel_get_attrs_count

LibIIO.CLibIIO.iio_channel_get_data

LibIIO.CLibIIO.iio_channel_get_data_format

LibIIO.CLibIIO.iio_channel_get_device

LibIIO.CLibIIO.iio_channel_get_id

LibIIO.CLibIIO.iio_channel_get_index

LibIIO.CLibIIO.iio_channel_get_modifier

LibIIO.CLibIIO.iio_channel_get_name

LibIIO.CLibIIO.iio_channel_get_type

LibIIO.CLibIIO.iio_channel_is_enabled

LibIIO.CLibIIO.iio_channel_is_output

LibIIO.CLibIIO.iio_channel_is_scan_element

LibIIO.CLibIIO.iio_channel_read

LibIIO.CLibIIO.iio_channel_read_raw

LibIIO.CLibIIO.iio_channel_set_data

LibIIO.CLibIIO.iio_channel_write

LibIIO.CLibIIO.iio_channel_write_raw

LibIIO.CLibIIO.iio_context_clone

LibIIO.CLibIIO.iio_context_destroy

LibIIO.CLibIIO.iio_context_find_device

LibIIO.CLibIIO.iio_context_get_attr

LibIIO.CLibIIO.iio_context_get_attr_value

LibIIO.CLibIIO.iio_context_get_attrs_count

LibIIO.CLibIIO.iio_context_get_description

LibIIO.CLibIIO.iio_context_get_device

LibIIO.CLibIIO.iio_context_get_devices_count

LibIIO.CLibIIO.iio_context_get_name

LibIIO.CLibIIO.iio_context_get_version

LibIIO.CLibIIO.iio_context_get_xml

LibIIO.CLibIIO.iio_context_info_get_description

LibIIO.CLibIIO.iio_context_info_get_uri

LibIIO.CLibIIO.iio_context_info_list_free

LibIIO.CLibIIO.iio_context_set_timeout

LibIIO.CLibIIO.iio_create_context_from_uri

LibIIO.CLibIIO.iio_create_default_context

LibIIO.CLibIIO.iio_create_local_context

LibIIO.CLibIIO.iio_create_network_context

LibIIO.CLibIIO.iio_create_scan_block

LibIIO.CLibIIO.iio_create_scan_context

LibIIO.CLibIIO.iio_create_xml_context

LibIIO.CLibIIO.iio_create_xml_context_mem

LibIIO.CLibIIO.iio_device_attr_read

LibIIO.CLibIIO.iio_device_attr_read_all

LibIIO.CLibIIO.iio_device_attr_read_bool

LibIIO.CLibIIO.iio_device_attr_read_double

LibIIO.CLibIIO.iio_device_attr_read_longlong

LibIIO.CLibIIO.iio_device_attr_write

LibIIO.CLibIIO.iio_device_attr_write_all

LibIIO.CLibIIO.iio_device_attr_write_bool

LibIIO.CLibIIO.iio_device_attr_write_double

LibIIO.CLibIIO.iio_device_attr_write_longlong

LibIIO.CLibIIO.iio_device_attr_write_raw

LibIIO.CLibIIO.iio_device_buffer_attr_read

LibIIO.CLibIIO.iio_device_buffer_attr_read_all

LibIIO.CLibIIO.iio_device_buffer_attr_read_bool

LibIIO.CLibIIO.iio_device_buffer_attr_read_double

LibIIO.CLibIIO.iio_device_buffer_attr_read_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write

LibIIO.CLibIIO.iio_device_buffer_attr_write_all

LibIIO.CLibIIO.iio_device_buffer_attr_write_bool

LibIIO.CLibIIO.iio_device_buffer_attr_write_double

LibIIO.CLibIIO.iio_device_buffer_attr_write_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write_raw

LibIIO.CLibIIO.iio_device_create_buffer

LibIIO.CLibIIO.iio_device_debug_attr_read

LibIIO.CLibIIO.iio_device_debug_attr_read_all

LibIIO.CLibIIO.iio_device_debug_attr_read_bool

LibIIO.CLibIIO.iio_device_debug_attr_read_double

LibIIO.CLibIIO.iio_device_debug_attr_read_longlong

LibIIO.CLibIIO.iio_device_debug_attr_write

LibIIO.CLibIIO.iio_device_debug_attr_write_all

LibIIO.CLibIIO.iio_device_debug_attr_write_bool

LibIIO.CLibIIO.iio_device_debug_attr_write_double

LibIIO.CLibIIO.iio_device_debug_attr_write_longlong

LibIIO.CLibIIO.iio_device_debug_attr_write_raw

LibIIO.CLibIIO.iio_device_find_attr

LibIIO.CLibIIO.iio_device_find_buffer_attr

LibIIO.CLibIIO.iio_device_find_channel

LibIIO.CLibIIO.iio_device_find_debug_attr

LibIIO.CLibIIO.iio_device_get_attr

LibIIO.CLibIIO.iio_device_get_attrs_count

LibIIO.CLibIIO.iio_device_get_buffer_attr

LibIIO.CLibIIO.iio_device_get_buffer_attrs_count

LibIIO.CLibIIO.iio_device_get_channel

LibIIO.CLibIIO.iio_device_get_channels_count

LibIIO.CLibIIO.iio_device_get_context

LibIIO.CLibIIO.iio_device_get_data

LibIIO.CLibIIO.iio_device_get_debug_attr

LibIIO.CLibIIO.iio_device_get_debug_attrs_count

LibIIO.CLibIIO.iio_device_get_id

LibIIO.CLibIIO.iio_device_get_label

LibIIO.CLibIIO.iio_device_get_name

LibIIO.CLibIIO.iio_device_get_sample_size

LibIIO.CLibIIO.iio_device_get_trigger

LibIIO.CLibIIO.iio_device_identify_filename

LibIIO.CLibIIO.iio_device_is_trigger

LibIIO.CLibIIO.iio_device_reg_read

LibIIO.CLibIIO.iio_device_reg_write

LibIIO.CLibIIO.iio_device_set_data

LibIIO.CLibIIO.iio_device_set_kernel_buffers_count

LibIIO.CLibIIO.iio_device_set_trigger

LibIIO.CLibIIO.iio_get_backend

LibIIO.CLibIIO.iio_get_backends_count

LibIIO.CLibIIO.iio_has_backend

LibIIO.CLibIIO.iio_library_get_version

LibIIO.CLibIIO.iio_scan_block_destroy

LibIIO.CLibIIO.iio_scan_block_get_info

LibIIO.CLibIIO.iio_scan_block_scan

LibIIO.CLibIIO.iio_scan_context_destroy

LibIIO.CLibIIO.iio_scan_context_get_info_list

LibIIO.CLibIIO.iio_strerror

LibIIO.attrs

LibIIO.attrs

LibIIO.attrs

LibIIO.buffer_attrs

LibIIO.cancel

LibIIO.channels

LibIIO.clone

LibIIO.context

LibIIO.data_format

LibIIO.debug_attrs

LibIIO.description

LibIIO.device

LibIIO.device

LibIIO.devices

LibIIO.enabled

LibIIO.enabled!

LibIIO.find_channel

LibIIO.find_device

LibIIO.frequency

LibIIO.frequency!

LibIIO.id

LibIIO.id

LibIIO.index

LibIIO.label

LibIIO.modifier

LibIIO.name

LibIIO.name

LibIIO.output

LibIIO.poll_fd

LibIIO.push

LibIIO.refill

LibIIO.reg_read

LibIIO.reg_write

LibIIO.sample_size

LibIIO.scan_contexts

LibIIO.scan_element

LibIIO.set_blocking_mode

LibIIO.set_kernel_buffers_count

LibIIO.set_timeout

LibIIO.step

LibIIO.trigger

LibIIO.trigger!

LibIIO.type

LibIIO.version

LibIIO.xml

LibIIO.Device — Type

Contains the representation of an IIO device.

LibIIO.Device — Method

Device(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Device instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this device

LibIIO.DeviceAttr — Type

DeviceAttr <: Attr

Represents an attribute of an IIO device

LibIIO.DeviceBufferAttr — Type

DeviceAttr <: Attr

Represents a buffer attribute of an IIO device

LibIIO.DeviceDebugAttr — Type

DeviceAttr <: Attr

Represents a debug attribute of an IIO device

LibIIO.Trigger — Type

Contains the representation of an IIO device that can act as a trigger.

LibIIO.Trigger — Method

Trigger(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Trigger instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this trigger

LibIIO.attrs — Method

attrs(d::AbstractDeviceOrTrigger)

List of attributes for the IIO device.

LibIIO.buffer_attrs — Method

buffer_attrs(d::AbstractDeviceOrTrigger)

List of buffer attributes for the IIO device.

LibIIO.channels — Method

channels(d::AbstractDeviceOrTrigger)

List of channels available with this IIO device.

LibIIO.context — Method

context(d::Device)

Context for the device.

LibIIO.debug_attrs — Method

debug_attrs(d::AbstractDeviceOrTrigger)

List of debug attributes for the IIO device.

LibIIO.find_channel — Function

find_channel(d::AbstractDeviceOrTrigger, name_or_id, is_output = false)

Find an IIO channel by its name or ID.

Parameters

d : The device instance

name_or_id : The name or ID of the channel to find

is_output : Set to true to search for an output channel

Returns

The IIO channel as Channel

LibIIO.frequency! — Method

frequency!(t::Trigger, value)

Set the trigger rate.

LibIIO.frequency — Method

frequency(t::Trigger)

Configured frequency (in Hz) of the trigger.

LibIIO.id — Method

id(d::AbstractDeviceOrTrigger)

An identifier of the device, only valid in this IIO context.

LibIIO.label — Method

label(d::AbstractDeviceOrTrigger)

The label of the device.

LibIIO.name — Method

name(d::AbstractDeviceOrTrigger)

The name of the device.

LibIIO.reg_read — Method

reg_read(d::AbstractDeviceOrTrigger, reg)

Read the content of a register of this device.

Parameters

d : The device instance

reg : The register address

Returns

The value of the register

LibIIO.reg_write — Method

reg_write(d::AbstractDeviceOrTrigger, reg, value)

Set a valie to one register of the device.

Parameters

d : The device instance

reg : The register address

value ; The value that will be used for this register

LibIIO.sample_size — Method

sample_size(d::AbstractDeviceOrTrigger)

Sample size of the device. The sample size varies each time channels get enabled or disabled.

LibIIO.set_kernel_buffers_count — Method

set_kernel_buffers_count(d::AbstractDeviceOrTrigger, count)

Set the number of kernel buffers to use with the specified device

Parameters

d : The device instance

count : The number of kernel buffers

LibIIO.trigger! — Method

trigger!(d::Device, trigger::Trigger)

Sets the configured trigger for this IIO device.

LibIIO.trigger — Method

trigger(d::Device)

Returns the configured trigger for this IIO device, if present in the current context.