LibIIO.CLibIIO.iio_buffer

LibIIO.CLibIIO.iio_chan_type

LibIIO.CLibIIO.iio_channel

LibIIO.CLibIIO.iio_context

LibIIO.CLibIIO.iio_context_info

LibIIO.CLibIIO.iio_data_format

LibIIO.CLibIIO.iio_device

LibIIO.CLibIIO.iio_modifier

LibIIO.CLibIIO.iio_buffer — Type

An input or output buffer, used to read or write samples

See libiio

LibIIO.CLibIIO.iio_chan_type — Type

iio_chan_type

IIO channel type

An IIO channel has a type specifying the type of data associated with the channel.

See libiio

LibIIO.CLibIIO.iio_channel — Type

Represents an input or output channel of a device

See libiio

LibIIO.CLibIIO.iio_context — Type

Contains the representation of an IIO context

See libiio

LibIIO.CLibIIO.iio_context_info — Type

The information related to a discovered context

See libiio

LibIIO.CLibIIO.iio_data_format — Type

iio_data_format

Contains the format of a data sample.

Data Fields

length::Cuint : Total length of the sample in bits

bits::Cuint : Length of the valuable data in the samples, in bits

shift:Cuint : Right-shift to apply when converting sample

is_signed::Cuchar : Contains true if the sample is signed

is_fully_defined::Cuchar : Contains true if the sample is fully defoned, sign extended, etc.

is_be::Cuchar : Contains true if the sample is in big-endian format

with_scale::Cuchar : Contains true if the sample should be scaled or converted

scale::Cdouble : Contains the scale to apply if with_scale is set

repeat::Cuint : Numver of times length repeats

See libiio

LibIIO.CLibIIO.iio_device — Type

Represents a device in the IIO context

See libiio

LibIIO.CLibIIO.iio_modifier — Type

iio_modifier

IIO channel type.

An IIO channel has a type specifying the type of data associated with the channel.

See libiio