LibIIO.Buffer

LibIIO.Buffer

LibIIO.CLibIIO.iio_buffer

LibIIO.CLibIIO.iio_chan_type

LibIIO.CLibIIO.iio_channel

LibIIO.CLibIIO.iio_context

LibIIO.CLibIIO.iio_context_info

LibIIO.CLibIIO.iio_data_format

LibIIO.CLibIIO.iio_device

LibIIO.CLibIIO.iio_modifier

LibIIO.Channel

LibIIO.Channel

LibIIO.ChannelAttr

LibIIO.ChannelAttr

LibIIO.Context

LibIIO.Context

LibIIO.Device

LibIIO.Device

LibIIO.DeviceAttr

LibIIO.DeviceBufferAttr

LibIIO.DeviceDebugAttr

LibIIO.LocalContext

LibIIO.LocalContext

LibIIO.NetworkContext

LibIIO.NetworkContext

LibIIO.Trigger

LibIIO.Trigger

LibIIO.XMLContext

LibIIO.XMLContext

Base.length

Base.read

Base.read

Base.write

Base.write

LibIIO.CLibIIO.iio_buffer_cancel

LibIIO.CLibIIO.iio_buffer_destroy

LibIIO.CLibIIO.iio_buffer_end

LibIIO.CLibIIO.iio_buffer_first

LibIIO.CLibIIO.iio_buffer_foreach_sample

LibIIO.CLibIIO.iio_buffer_get_data

LibIIO.CLibIIO.iio_buffer_get_device

LibIIO.CLibIIO.iio_buffer_get_poll_fd

LibIIO.CLibIIO.iio_buffer_push

LibIIO.CLibIIO.iio_buffer_push_partial

LibIIO.CLibIIO.iio_buffer_refill

LibIIO.CLibIIO.iio_buffer_set_blocking_mode

LibIIO.CLibIIO.iio_buffer_set_data

LibIIO.CLibIIO.iio_buffer_start

LibIIO.CLibIIO.iio_buffer_step

LibIIO.CLibIIO.iio_channel_attr_get_filename

LibIIO.CLibIIO.iio_channel_attr_read

LibIIO.CLibIIO.iio_channel_attr_read_all

LibIIO.CLibIIO.iio_channel_attr_read_bool

LibIIO.CLibIIO.iio_channel_attr_read_double

LibIIO.CLibIIO.iio_channel_attr_read_longlong

LibIIO.CLibIIO.iio_channel_attr_write

LibIIO.CLibIIO.iio_channel_attr_write_all

LibIIO.CLibIIO.iio_channel_attr_write_bool

LibIIO.CLibIIO.iio_channel_attr_write_double

LibIIO.CLibIIO.iio_channel_attr_write_longlong

LibIIO.CLibIIO.iio_channel_attr_write_raw

LibIIO.CLibIIO.iio_channel_convert

LibIIO.CLibIIO.iio_channel_convert_inverse

LibIIO.CLibIIO.iio_channel_disable

LibIIO.CLibIIO.iio_channel_enable

LibIIO.CLibIIO.iio_channel_find_attr

LibIIO.CLibIIO.iio_channel_get_attr

LibIIO.CLibIIO.iio_channel_get_attrs_count

LibIIO.CLibIIO.iio_channel_get_data

LibIIO.CLibIIO.iio_channel_get_data_format

LibIIO.CLibIIO.iio_channel_get_device

LibIIO.CLibIIO.iio_channel_get_id

LibIIO.CLibIIO.iio_channel_get_index

LibIIO.CLibIIO.iio_channel_get_modifier

LibIIO.CLibIIO.iio_channel_get_name

LibIIO.CLibIIO.iio_channel_get_type

LibIIO.CLibIIO.iio_channel_is_enabled

LibIIO.CLibIIO.iio_channel_is_output

LibIIO.CLibIIO.iio_channel_is_scan_element

LibIIO.CLibIIO.iio_channel_read

LibIIO.CLibIIO.iio_channel_read_raw

LibIIO.CLibIIO.iio_channel_set_data

LibIIO.CLibIIO.iio_channel_write

LibIIO.CLibIIO.iio_channel_write_raw

LibIIO.CLibIIO.iio_context_clone

LibIIO.CLibIIO.iio_context_destroy

LibIIO.CLibIIO.iio_context_find_device

LibIIO.CLibIIO.iio_context_get_attr

LibIIO.CLibIIO.iio_context_get_attr_value

LibIIO.CLibIIO.iio_context_get_attrs_count

LibIIO.CLibIIO.iio_context_get_description

LibIIO.CLibIIO.iio_context_get_device

LibIIO.CLibIIO.iio_context_get_devices_count

LibIIO.CLibIIO.iio_context_get_name

LibIIO.CLibIIO.iio_context_get_version

LibIIO.CLibIIO.iio_context_get_xml

LibIIO.CLibIIO.iio_context_info_get_description

LibIIO.CLibIIO.iio_context_info_get_uri

LibIIO.CLibIIO.iio_context_info_list_free

LibIIO.CLibIIO.iio_context_set_timeout

LibIIO.CLibIIO.iio_create_context_from_uri

LibIIO.CLibIIO.iio_create_default_context

LibIIO.CLibIIO.iio_create_local_context

LibIIO.CLibIIO.iio_create_network_context

LibIIO.CLibIIO.iio_create_scan_block

LibIIO.CLibIIO.iio_create_scan_context

LibIIO.CLibIIO.iio_create_xml_context

LibIIO.CLibIIO.iio_create_xml_context_mem

LibIIO.CLibIIO.iio_device_attr_read

LibIIO.CLibIIO.iio_device_attr_read_all

LibIIO.CLibIIO.iio_device_attr_read_bool

LibIIO.CLibIIO.iio_device_attr_read_double

LibIIO.CLibIIO.iio_device_attr_read_longlong

LibIIO.CLibIIO.iio_device_attr_write

LibIIO.CLibIIO.iio_device_attr_write_all

LibIIO.CLibIIO.iio_device_attr_write_bool

LibIIO.CLibIIO.iio_device_attr_write_double

LibIIO.CLibIIO.iio_device_attr_write_longlong

LibIIO.CLibIIO.iio_device_attr_write_raw

LibIIO.CLibIIO.iio_device_buffer_attr_read

LibIIO.CLibIIO.iio_device_buffer_attr_read_all

LibIIO.CLibIIO.iio_device_buffer_attr_read_bool

LibIIO.CLibIIO.iio_device_buffer_attr_read_double

LibIIO.CLibIIO.iio_device_buffer_attr_read_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write

LibIIO.CLibIIO.iio_device_buffer_attr_write_all

LibIIO.CLibIIO.iio_device_buffer_attr_write_bool

LibIIO.CLibIIO.iio_device_buffer_attr_write_double

LibIIO.CLibIIO.iio_device_buffer_attr_write_longlong

LibIIO.CLibIIO.iio_device_buffer_attr_write_raw

LibIIO.CLibIIO.iio_device_create_buffer

LibIIO.CLibIIO.iio_device_debug_attr_read

LibIIO.CLibIIO.iio_device_debug_attr_read_all

LibIIO.CLibIIO.iio_device_debug_attr_read_bool

LibIIO.CLibIIO.iio_device_debug_attr_read_double

LibIIO.CLibIIO.iio_device_debug_attr_read_longlong

LibIIO.CLibIIO.iio_device_debug_attr_write

LibIIO.CLibIIO.iio_device_debug_attr_write_all

LibIIO.CLibIIO.iio_device_debug_attr_write_bool

LibIIO.CLibIIO.iio_device_debug_attr_write_double

LibIIO.CLibIIO.iio_device_debug_attr_write_longlong

LibIIO.CLibIIO.iio_device_debug_attr_write_raw

LibIIO.CLibIIO.iio_device_find_attr

LibIIO.CLibIIO.iio_device_find_buffer_attr

LibIIO.CLibIIO.iio_device_find_channel

LibIIO.CLibIIO.iio_device_find_debug_attr

LibIIO.CLibIIO.iio_device_get_attr

LibIIO.CLibIIO.iio_device_get_attrs_count

LibIIO.CLibIIO.iio_device_get_buffer_attr

LibIIO.CLibIIO.iio_device_get_buffer_attrs_count

LibIIO.CLibIIO.iio_device_get_channel

LibIIO.CLibIIO.iio_device_get_channels_count

LibIIO.CLibIIO.iio_device_get_context

LibIIO.CLibIIO.iio_device_get_data

LibIIO.CLibIIO.iio_device_get_debug_attr

LibIIO.CLibIIO.iio_device_get_debug_attrs_count

LibIIO.CLibIIO.iio_device_get_id

LibIIO.CLibIIO.iio_device_get_label

LibIIO.CLibIIO.iio_device_get_name

LibIIO.CLibIIO.iio_device_get_sample_size

LibIIO.CLibIIO.iio_device_get_trigger

LibIIO.CLibIIO.iio_device_identify_filename

LibIIO.CLibIIO.iio_device_is_trigger

LibIIO.CLibIIO.iio_device_reg_read

LibIIO.CLibIIO.iio_device_reg_write

LibIIO.CLibIIO.iio_device_set_data

LibIIO.CLibIIO.iio_device_set_kernel_buffers_count

LibIIO.CLibIIO.iio_device_set_trigger

LibIIO.CLibIIO.iio_get_backend

LibIIO.CLibIIO.iio_get_backends_count

LibIIO.CLibIIO.iio_has_backend

LibIIO.CLibIIO.iio_library_get_version

LibIIO.CLibIIO.iio_scan_block_destroy

LibIIO.CLibIIO.iio_scan_block_get_info

LibIIO.CLibIIO.iio_scan_block_scan

LibIIO.CLibIIO.iio_scan_context_destroy

LibIIO.CLibIIO.iio_scan_context_get_info_list

LibIIO.CLibIIO.iio_strerror

LibIIO.attrs

LibIIO.attrs

LibIIO.attrs

LibIIO.buffer_attrs

LibIIO.cancel

LibIIO.channels

LibIIO.clone

LibIIO.context

LibIIO.data_format

LibIIO.debug_attrs

LibIIO.description

LibIIO.device

LibIIO.device

LibIIO.devices

LibIIO.enabled

LibIIO.enabled!

LibIIO.find_channel

LibIIO.find_device

LibIIO.frequency

LibIIO.frequency!

LibIIO.id

LibIIO.id

LibIIO.index

LibIIO.label

LibIIO.modifier

LibIIO.name

LibIIO.name

LibIIO.output

LibIIO.poll_fd

LibIIO.push

LibIIO.refill

LibIIO.reg_read

LibIIO.reg_write

LibIIO.sample_size

LibIIO.scan_contexts

LibIIO.scan_element

LibIIO.set_blocking_mode

LibIIO.set_kernel_buffers_count

LibIIO.set_timeout

LibIIO.step

LibIIO.trigger

LibIIO.trigger!

LibIIO.type

LibIIO.version

LibIIO.xml

LibIIO.CLibIIO.iio_context_info_get_description — Method

iio_context_info_get_description(info)

Get a description of a discovered context.

Parameters

Ptr{iio_context_info}: A pointer to a iio_context_info structure.

Returns

A String containing the description.

See libiio

LibIIO.CLibIIO.iio_context_info_get_uri — Method

iio_context_info_get_uri(info)

Get the URI of a discovered context.

Parameters

Ptr{iio_context_info}: A pointer to a iio_context_info structure.

Returns

A String containing the URI.

See libiio

LibIIO.CLibIIO.iio_context_info_list_free — Method

iio_context_info_list_free(info)

Free a context info list.

Parameters

info::Ptr{Ptr{iio_context_info}} : A pointer to a iio_context_info pointer.

See libiio

LibIIO.CLibIIO.iio_create_scan_block — Function

iio_create_scan_block(backend, flags = 0)

Parameters

backend::String : A string containing the backend to use for scanning. Can be empty to use all available backends.

flags::Cuint (optional) : Unused for now. Set to 0.

Returns

on success, a pointer to an iio_scan_block structure

on failure, an error is raised

See libiio

LibIIO.CLibIIO.iio_create_scan_context — Function

iio_create_scan_context(backend, flags = 0)

Create a scan context.

Parameters

backend::String : A string containing a comma-separated list of the backend(s) to use for scanning.

flags::Cuint : Unused for now. Set to 0.

Returns

on success, a pointer to a iio_scan_context structure

On failure, an error is raised

Note

Libiio version 0.20 and above can handle multiple strings, for instance "local:usb:", "ip:usb:", "local:usb:ip:", and require a colon as the delimiter. Libiio version 0.24 and above prefer a comma instead of colon as the delimiter, and handle specifying backend-specific information. For instance, "local,usb=0456:*" will scan the local backend and limit scans on USB to vendor ID 0x0456, and accept all product IDs. The "usb=0456:b673" string would limit the scan to the device with this particular VID/PID. Both IDs are expected in hexadecimal, no 0x prefix needed.

LibIIO.CLibIIO.iio_scan_block_destroy — Method

iio_scan_block_destroy(blk)

Destroy the given scan block.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

Note

After that function, the iioscanblock pointer shall be invalid.

Introduced in version 0.20.

See libiio

LibIIO.CLibIIO.iio_scan_block_get_info — Method

iio_scan_block_get_info(blk, index)

Get the iio_context_info for a particular context.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

index::Cuint : The index corresponding to the context.

Returns

A pointer to the iio_context_info for the context

On success, a pointer to the specified iio_context_info

On failure, an error is raised

LibIIO.CLibIIO.iio_scan_block_scan — Method

iio_scan_block_scan(blk)

Enumerate available contexts via scan block.

Parameters

blk::Ptr{iio_scan_block} : A pointer to an iio_scan_block structure

Returns

On success, the number of contexts found.

On failure, a negative error number.

See libiio

LibIIO.CLibIIO.iio_scan_context_destroy — Method

iio_scan_context_destroy(ctx)

Destroy the given scan context.

Parameters

ctx::Ptr{iio_scan_context} : A pointer to an iio_scan_context structure

Note

After that function, the iio_scan_context pointer shall be invalid.

See libiio

LibIIO.CLibIIO.iio_scan_context_get_info_list — Method

iio_scan_context_get_info_list(ctx, info)

Enumerate available contexts.

Parameters

ctx::Ptr{iio_scan_context} : A pointer to an iio_scan_context structure

info::Ref{Ptr{Ptr{iio_context_info}}} : A pointer to a list of iio_context_info pointers. The pointed variable will be initialized on success.

Returns

On success, the number of contexts found.

On failure, a negative error number.

See libiio