LibIIO.Device

LibIIO.Device

LibIIO.DeviceAttr

LibIIO.DeviceBufferAttr

LibIIO.DeviceDebugAttr

LibIIO.Trigger

LibIIO.Trigger

LibIIO.attrs

LibIIO.buffer_attrs

LibIIO.channels

LibIIO.context

LibIIO.debug_attrs

LibIIO.find_channel

LibIIO.frequency

LibIIO.frequency!

LibIIO.id

LibIIO.label

LibIIO.name

LibIIO.reg_read

LibIIO.reg_write

LibIIO.sample_size

LibIIO.set_kernel_buffers_count

LibIIO.trigger

LibIIO.trigger!

LibIIO.Device — Type

Contains the representation of an IIO device.

LibIIO.Device — Method

Device(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Device instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this device

LibIIO.DeviceAttr — Type

DeviceAttr <: Attr

Represents an attribute of an IIO device

LibIIO.DeviceBufferAttr — Type

DeviceAttr <: Attr

Represents a buffer attribute of an IIO device

LibIIO.DeviceDebugAttr — Type

DeviceAttr <: Attr

Represents a debug attribute of an IIO device

LibIIO.Trigger — Type

Contains the representation of an IIO device that can act as a trigger.

LibIIO.Trigger — Method

Trigger(ctx::AbstractContext, device::Ptr{iio_device})

Initializes a new Trigger instance.

Parameters

ctx : The IIO context instance with which the device is accessed

device : A pointer to an iio_device which represents this trigger

LibIIO.attrs — Method

attrs(d::AbstractDeviceOrTrigger)

List of attributes for the IIO device.

LibIIO.buffer_attrs — Method

buffer_attrs(d::AbstractDeviceOrTrigger)

List of buffer attributes for the IIO device.

LibIIO.channels — Method

channels(d::AbstractDeviceOrTrigger)

List of channels available with this IIO device.

LibIIO.context — Method

context(d::Device)

Context for the device.

LibIIO.debug_attrs — Method

debug_attrs(d::AbstractDeviceOrTrigger)

List of debug attributes for the IIO device.

LibIIO.find_channel — Function

find_channel(d::AbstractDeviceOrTrigger, name_or_id, is_output = false)

Find an IIO channel by its name or ID.

Parameters

d : The device instance

name_or_id : The name or ID of the channel to find

is_output : Set to true to search for an output channel

Returns

The IIO channel as Channel

LibIIO.frequency! — Method

frequency!(t::Trigger, value)

Set the trigger rate.

LibIIO.frequency — Method

frequency(t::Trigger)

Configured frequency (in Hz) of the trigger.

LibIIO.id — Method

id(d::AbstractDeviceOrTrigger)

An identifier of the device, only valid in this IIO context.

LibIIO.label — Method

label(d::AbstractDeviceOrTrigger)

The label of the device.

LibIIO.name — Method

name(d::AbstractDeviceOrTrigger)

The name of the device.

LibIIO.reg_read — Method

reg_read(d::AbstractDeviceOrTrigger, reg)

Read the content of a register of this device.

Parameters

d : The device instance

reg : The register address

Returns

The value of the register

LibIIO.reg_write — Method

reg_write(d::AbstractDeviceOrTrigger, reg, value)

Set a valie to one register of the device.

Parameters

d : The device instance

reg : The register address

value ; The value that will be used for this register

LibIIO.sample_size — Method

sample_size(d::AbstractDeviceOrTrigger)

Sample size of the device. The sample size varies each time channels get enabled or disabled.

LibIIO.set_kernel_buffers_count — Method

set_kernel_buffers_count(d::AbstractDeviceOrTrigger, count)

Set the number of kernel buffers to use with the specified device

Parameters

d : The device instance

count : The number of kernel buffers

LibIIO.trigger! — Method

trigger!(d::Device, trigger::Trigger)

Sets the configured trigger for this IIO device.

LibIIO.trigger — Method

trigger(d::Device)

Returns the configured trigger for this IIO device, if present in the current context.